+++ src/dashboard/CHANGELOG.md
@@ -20,6 +20,8 @@ and this project adheres to
 - add consent form to manage consents of one or many entities 
 - add admin integration for Entity, DeliveryPoint and Consent
 - add mass admin action (make revoked and make awaiting) for consents
+- block the updates of all new data if a consent has the status `VALIDATED`
+- block the deletion of consent if it has the status `VALIDATED`
 - integration of custom 403, 404 and 500 pages 
 - sentry integration
 - added a signal on the creation of a delivery point. This signal allows the creation